+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortVirtualTool.java
@@ -1,9 +1,7 @@
 package ai.mcpdirect.backend.dao.entity.aitool;
 
 public class AIPortVirtualTool extends AIPortTool{
-    public long toolId;
-    public String tags;
-
+    public long originalToolId;
     public AIPortVirtualTool() {}
 
 }